// Oraculos.ts-main/src/types.ts
export type OracleType =
  | 'tarot'
  | 'cigano'
  | 'astrologia'
  | 'numerologia'
  | 'buzios'
  | 'ifa'
  | 'runas'
  | 'iching'
  | 'cristais'
  | 'mesaradionica';